Extract timeline to a component and reuse for celestial body spacecraft visits
