+++ src/components/FactSheet/SpacecraftVisits.tsx
@@ -1,8 +1,11 @@
-import { Stack, Title } from '@mantine/core';
+import { Box, Stack, Title } from '@mantine/core';
+import { ReactNode, useState } from 'react';
 import { useFactSheetPadding } from '../../hooks/useFactSheetPadding.ts';
 import { UpdateSettings } from '../../lib/state.ts';
 import { CelestialBody, Spacecraft } from '../../lib/types.ts';
+import { DEFAULT_SPACECRAFT_COLOR } from '../../lib/utils.ts';
 import { SpacecraftCard } from './Spacecraft/SpacecraftCard.tsx';
+import { Timeline } from './Timeline.tsx';
 
 type Props = {
   spacecraft: Array<Spacecraft>;
@@ -12,17 +15,23 @@ type Props = {
 };
 export function SpacecraftVisits({ spacecraft, body, updateSettings, title = 'Spacecraft Visits' }: Props) {
   const padding = useFactSheetPadding();
+  const [activeIndex, setActiveIndex] = useState<number | undefined>();
+
+  const TimelineItems = spacecraft.map<[Date, ReactNode]>((s, i) => [
+    s.start,
+    <Box key={`${s.name}-${i}`} onMouseEnter={() => setActiveIndex(i)} onMouseLeave={() => setActiveIndex(undefined)}>
+      <SpacecraftCard spacecraft={s} body={body} onClick={() => updateSettings({ center: s.id, hover: null })} />
+    </Box>,
+  ]);
+
   return spacecraft.length > 0 ? (
     <Stack gap="xs" {...padding}>
       <Title order={5}>{title}</Title>
-      {spacecraft.map((s, i) => (
-        <SpacecraftCard
-          key={`${s.name}-${i}`}
-          spacecraft={s}
-          body={body}
-          onClick={() => updateSettings({ center: s.id, hover: null })}
-        />
-      ))}
+      <Timeline
+        datedItems={TimelineItems}
+        activeIndex={activeIndex}
+        accentColor={body?.style?.fgColor ?? DEFAULT_SPACECRAFT_COLOR}
+      />
     </Stack>
   ) : (
     <></>
